+++ odbc/odbc_handles.h
@@ -8,6 +8,7 @@
 #include <vector>
 #include <map>
 #include <mutex>
+#include <optional>
 
 namespace rs::odbc {
 
@@ -188,6 +189,10 @@ public:
   // Metadata functions
   SQLRETURN get_num_result_cols(SQLSMALLINT* column_count);
   SQLRETURN get_type_info(SQLSMALLINT data_type);
+  SQLRETURN tables(const std::optional<std::string>& catalog_name,
+                   const std::optional<std::string>& schema_name,
+                   const std::optional<std::string>& table_name,
+                   const std::optional<std::string>& table_type);
   SQLRETURN row_count(SQLLEN* row_count);
   SQLRETURN describe_col(SQLUSMALLINT column_number, SQLCHAR* column_name, SQLSMALLINT name_buffer_length,
                         SQLSMALLINT* name_length, SQLSMALLINT* data_type, SQLULEN* column_size,