+++ tests/unit/test_attribute_apis.cpp
@@ -158,7 +158,7 @@ TEST_F(AttributeApisTest, ReportsImplementedFunctions) {
       connection_, SQL_API_SQLDRIVERCONNECT, &supported));
   EXPECT_EQ(SQL_TRUE, supported);
   EXPECT_EQ(SQL_SUCCESS, SQLGetFunctions(
-      connection_, SQL_API_SQLTABLES, &supported));
+      connection_, SQL_API_SQLCOLUMNS, &supported));
   EXPECT_EQ(SQL_FALSE, supported);
 
   SQLUSMALLINT odbc2_functions[100]{};
@@ -166,7 +166,7 @@ TEST_F(AttributeApisTest, ReportsImplementedFunctions) {
       connection_, SQL_API_ALL_FUNCTIONS, odbc2_functions));
   EXPECT_EQ(SQL_TRUE, odbc2_functions[SQL_API_SQLCONNECT]);
   EXPECT_EQ(SQL_TRUE, odbc2_functions[SQL_API_SQLDRIVERCONNECT]);
-  EXPECT_EQ(SQL_FALSE, odbc2_functions[SQL_API_SQLTABLES]);
+  EXPECT_EQ(SQL_FALSE, odbc2_functions[SQL_API_SQLCOLUMNS]);
 
   SQLUSMALLINT odbc3_functions[SQL_API_ODBC3_ALL_FUNCTIONS_SIZE]{};
   EXPECT_EQ(SQL_SUCCESS, SQLGetFunctions(
@@ -176,7 +176,7 @@ TEST_F(AttributeApisTest, ReportsImplementedFunctions) {
   EXPECT_EQ(SQL_TRUE,
             SQL_FUNC_EXISTS(odbc3_functions, SQL_API_SQLDRIVERCONNECT));
   EXPECT_EQ(SQL_FALSE,
-            SQL_FUNC_EXISTS(odbc3_functions, SQL_API_SQLTABLES));
+            SQL_FUNC_EXISTS(odbc3_functions, SQL_API_SQLCOLUMNS));
 
   EXPECT_EQ(SQL_ERROR, SQLGetFunctions(
       connection_, SQL_API_SQLCONNECT, nullptr));